Finalizado

+++ src/main/java/com/garra/academico/entities/Categoria.java
@@ -18,7 +18,7 @@ public class Categoria {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	
 	@Column(columnDefinition = "TEXT")
 	private String descricao;
@@ -29,16 +29,16 @@ public class Categoria {
 	public Categoria() {
 	}
 
-	public Categoria(Long id, String descricao) {
+	public Categoria(Integer id, String descricao) {
 		this.id = id;
 		this.descricao = descricao;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 
@@ -53,6 +53,8 @@ public class Categoria {
 	public List<Atividade> getAtividades() {
 		return atividades;
 	}
+	
+	
 
 	@Override
 	public int hashCode() {
